refactor: Remove broken V1 BlockSparse implementation, make V2 the default
The original BlockSparse implementation had a fundamental flaw where it
allocated full dense attention matrices, completely defeating the purpose
of sparse attention. This made it use MORE memory than regular attention.

Changes:
- Remove broken V1 BlockSparseRingDilatedAttention implementation
- Remove broken V1 BlockSparseRingMultiheadDilatedAttention implementation
- Rename V2 implementations to be the main BlockSparse implementations
- Update all imports and references to use the memory-efficient version
- Update benchmark configurations to use more appropriate sparsity (90% sparse)
- Remove V2-specific test files and documentation

The new default BlockSparse implementation:
- Never materializes full attention matrices
- Uses 83-96% less memory than the old version
- Runs 8-126x faster depending on sequence length
- Returns sparse attention weights in COO format

This is a breaking change, but justified since the V1 implementation
was fundamentally broken and misleading to users.
