+++ dilated_attention_pytorch/block_sparse_ring_dilated_attention.py
@@ -1,682 +1,77 @@
 """
-Block-Sparse Ring Dilated Attention Implementation
+Block-Sparse Ring Dilated Attention - Memory Efficient Implementation
 
-This module combines block-sparse attention patterns with Ring Attention for
-maximum efficiency and scalability. It provides 5-50x speedup over dense attention
-while maintaining 95-99% quality retention.
-
-Key Features:
-- Multiple sparsity patterns (local window, dilated sparse, content-adaptive)
-- Seamless integration with existing Ring Attention infrastructure
-- Hardware-optimized kernels for H100/MI300X
-- Dynamic pattern adaptation based on content
-- Production-ready with comprehensive error handling
-
-Performance Benefits:
-- 5-50x speedup depending on sparsity level
-- 75-95% memory reduction
-- 75% communication bandwidth savings
-- Near-perfect quality retention (95-99%)
+This implementation truly leverages sparsity for memory efficiency by:
+1. Never materializing full attention matrices
+2. Processing only active blocks
+3. Using sparse tensor representations where possible
+4. Returning sparse attention weights in COO format
 """
 
-import gc
 import math
 import threading
-import time
-import warnings
-from collections import OrderedDict
-from collections.abc import Sequence
 from dataclasses import dataclass
-from typing import Any
 
 import torch
-import torch.distributed as dist
-import torch.nn.functional as F
-from torch import Tensor, nn
+from torch import Tensor
 
-# Import base Ring Attention implementation
 from .ring_dilated_attention import RingDilatedAttention
-
-# Handle torch.nn.attention availability for older PyTorch versions
-try:
-    from torch.nn.attention import SDPBackend
-
-    HAS_SDPA_KERNEL = True
-except ImportError:
-    HAS_SDPA_KERNEL = False
-
-    class SDPBackend:
-        FLASH_ATTENTION = "flash_attention"
-        EFFICIENT_ATTENTION = "efficient_attention"
-        MATH = "math"
-
-
-# Optional imports
-try:
-    from flash_attn import flash_attn_func
-
-    HAS_FLASH_ATTN = True
-except ImportError:
-    HAS_FLASH_ATTN = False
-
-
-class BlockSparseMemoryPool:
-    """
-    Advanced memory pool with hot cache and adaptive cleanup.
-
-    Features:
-    - Hot key caching for frequently accessed buffers
-    - Adaptive cleanup based on GPU memory pressure
-    - Thread-safe operations
-    - Smart buffer reuse with resize optimization
-    """
-
-    def __init__(self, max_pool_size: int = 100, hot_cache_size: int = 10):
-        self.max_pool_size = max_pool_size
-        self.hot_cache_size = hot_cache_size
-        self.pool: dict[tuple, list[Tensor]] = {}
-        self.hot_cache: OrderedDict[tuple, Tensor] = OrderedDict()
-        self.access_counts: dict[tuple, int] = {}
-        self._lock = threading.Lock()
-
-        # Adaptive thresholds
-        self.cleanup_threshold = 0.5  # Start conservative
-        self.last_cleanup_time = time.time()
-        self.cleanup_interval = 60.0  # seconds
-
-    def get_buffer(
-        self, shape: tuple[int, ...], dtype: torch.dtype, device: torch.device
-    ) -> Tensor:
-        """Get buffer from pool with hot cache check."""
-        key = (shape, dtype, device)
-
-        with self._lock:
-            # Check hot cache first
-            if key in self.hot_cache:
-                buffer = self.hot_cache.pop(key)
-                self.hot_cache[key] = buffer  # Move to end (LRU)
-                self.access_counts[key] = self.access_counts.get(key, 0) + 1
-                return buffer
-
-            # Check regular pool
-            if self.pool.get(key):
-                buffer = self.pool[key].pop()
-                self._update_hot_cache(key, buffer)
-                return buffer
-
-            # Create new buffer
-            try:
-                buffer = torch.empty(shape, dtype=dtype, device=device)
-                self._update_hot_cache(key, buffer)
-                return buffer
-            except torch.cuda.OutOfMemoryError:
-                self._emergency_cleanup()
-                # Retry after cleanup
-                return torch.empty(shape, dtype=dtype, device=device)
-
-    def return_buffer(self, buffer: Tensor):
-        """Return buffer to pool with smart placement."""
-        if buffer is None:
-            return
-
-        key = (buffer.shape, buffer.dtype, buffer.device)
-
-        with self._lock:
-            # Check if it should go to hot cache
-            access_count = self.access_counts.get(key, 0)
-            if access_count > 5 and len(self.hot_cache) < self.hot_cache_size:
-                self.hot_cache[key] = buffer
-            else:
-                # Add to regular pool with global size limit
-                if key not in self.pool:
-                    self.pool[key] = []
-
-                # Check global pool size
-                total_buffers = sum(len(buffers) for buffers in self.pool.values())
-                if total_buffers >= self.max_pool_size:
-                    # Evict from pool with most buffers
-                    largest_key = max(self.pool.keys(), key=lambda k: len(self.pool[k]))
-                    if self.pool[largest_key]:
-                        self.pool[largest_key].pop(0)  # Remove oldest
-
-                self.pool[key].append(buffer)
-
-            # Periodic cleanup
-            if time.time() - self.last_cleanup_time > self.cleanup_interval:
-                self._adaptive_cleanup()
-
-    def _update_hot_cache(self, key: tuple, buffer: Tensor):
-        """Update hot cache with LRU eviction."""
-        self.access_counts[key] = self.access_counts.get(key, 0) + 1
-
-        if self.access_counts[key] > 5:  # Threshold for hot
-            if len(self.hot_cache) >= self.hot_cache_size:
-                # Evict least recently used
-                evicted_key, evicted_buffer = self.hot_cache.popitem(last=False)
-                # Return evicted buffer to regular pool
-                if evicted_key not in self.pool:
-                    self.pool[evicted_key] = []
-                if len(self.pool[evicted_key]) < self.max_pool_size:
-                    self.pool[evicted_key].append(evicted_buffer)
-
-            self.hot_cache[key] = buffer
-
-    def _adaptive_cleanup(self):
-        """Adaptive cleanup based on memory pressure."""
-        if torch.cuda.is_available():
-            free_memory = torch.cuda.mem_get_info()[0]
-            total_memory = torch.cuda.mem_get_info()[1]
-            memory_usage = 1.0 - (free_memory / total_memory)
-
-            # Adjust cleanup threshold based on memory pressure
-            if memory_usage > 0.9:
-                self.cleanup_threshold = 0.1  # Aggressive cleanup
-            elif memory_usage > 0.7:
-                self.cleanup_threshold = 0.3  # Moderate cleanup
-            else:
-                self.cleanup_threshold = 0.5  # Conservative cleanup
-
-        # Clean up pools
-        for key in list(self.pool.keys()):
-            if key in self.pool:
-                keep_count = int(len(self.pool[key]) * self.cleanup_threshold)
-                self.pool[key] = self.pool[key][:keep_count]
-
-        # Clean up access counts for keys not in use
-        active_keys = set(self.pool.keys()) | set(self.hot_cache.keys())
-        self.access_counts = {k: v for k, v in self.access_counts.items() if k in active_keys}
-
-        self.last_cleanup_time = time.time()
-
-    def _emergency_cleanup(self):
-        """Emergency cleanup on OOM."""
-        with self._lock:
-            # Clear everything except hot cache
-            self.pool.clear()
-            # Reduce hot cache
-            while len(self.hot_cache) > self.hot_cache_size // 2:
-                self.hot_cache.popitem(last=False)
-            gc.collect()
-            if torch.cuda.is_available():
-                torch.cuda.empty_cache()
 
 
 @dataclass
 class SparsePatternConfig:
     """Configuration for sparse attention patterns"""
 
-    pattern_type: str = (
-        "dilated_sparse"  # 'local_window', 'dilated_sparse', 'global_local', 'adaptive'
-    )
-    sparsity_ratio: float = 0.25  # Fraction of blocks to compute (0.1 = 90% sparse)
+    pattern_type: str = "dilated_sparse"  # 'local_window', 'dilated_sparse', 'global_local'
+    sparsity_ratio: float = 0.1  # Fraction of blocks to compute (0.1 = 90% sparse)
     block_size: int = 128  # Tokens per block
     local_window_size: int = 512  # For local window patterns
     global_tokens: int = 64  # For global+local patterns
-    adaptation_rate: float = 0.1  # For adaptive patterns
-    min_sparsity: float = 0.05  # Minimum sparsity to maintain
-    max_sparsity: float = 0.95  # Maximum sparsity to maintain
-
-    def __post_init__(self):
-        """Validate configuration parameters."""
-        if self.block_size <= 0:
-            raise ValueError(f"block_size must be positive, got {self.block_size}")
-
-
-class SparsePatternGenerator:
-    """Utility class for generating various sparse attention patterns"""
-
-    def __init__(self, config: SparsePatternConfig, max_cache_size: int = 50):
-        self.config = config
-        self.max_cache_size = max_cache_size
-        self.pattern_cache: dict[tuple, torch.Tensor] = {}
-        self.adaptive_history: list[torch.Tensor] = []
-        self._cache_lock = threading.Lock()
-        self._cache_access_order = []  # For LRU eviction
-
-    def create_pattern(
-        self, seq_len: int, num_heads: int = 1, device: torch.device = None
-    ) -> torch.Tensor:
-        """Create sparsity pattern based on configuration"""
-        if device is None:
-            device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-
-        cache_key = (
-            seq_len,
-            num_heads,
-            self.config.pattern_type,
-            self.config.sparsity_ratio,
-        )
-
-        with self._cache_lock:
-            if cache_key in self.pattern_cache:
-                return self.pattern_cache[cache_key].to(device)
-
-        num_blocks = seq_len // self.config.block_size
-
-        if self.config.pattern_type == "local_window":
-            pattern = self._create_local_window_pattern(num_blocks, device)
-        elif self.config.pattern_type == "dilated_sparse":
-            pattern = self._create_dilated_sparse_pattern(num_blocks, device)
-        elif self.config.pattern_type == "global_local":
-            pattern = self._create_global_local_pattern(num_blocks, device)
-        elif self.config.pattern_type == "adaptive":
-            pattern = self._create_adaptive_pattern(num_blocks, device)
-        else:
-            raise ValueError(f"Unknown pattern type: {self.config.pattern_type}")
-
-        # Ensure target sparsity
-        pattern = self._enforce_target_sparsity(pattern, self.config.sparsity_ratio)
-
-        with self._cache_lock:
-            # Enforce cache size limit
-            if len(self.pattern_cache) >= self.max_cache_size:
-                # Evict LRU entry
-                if self._cache_access_order:
-                    lru_key = self._cache_access_order.pop(0)
-                    if lru_key in self.pattern_cache:
-                        del self.pattern_cache[lru_key]
-
-            self.pattern_cache[cache_key] = pattern.cpu()
-
-            # Update access order
-            if cache_key in self._cache_access_order:
-                self._cache_access_order.remove(cache_key)
-            self._cache_access_order.append(cache_key)
-
-        return pattern.to(device)
-
-    def _create_local_window_pattern(self, num_blocks: int, device: torch.device) -> torch.Tensor:
-        """Create local window sparsity pattern"""
-        pattern = torch.zeros(num_blocks, num_blocks, dtype=torch.bool, device=device)
-        window_blocks = self.config.local_window_size // self.config.block_size
-
-        for i in range(num_blocks):
-            start = max(0, i - window_blocks // 2)
-            end = min(num_blocks, i + window_blocks // 2 + 1)
-            pattern[i, start:end] = True
-
-        return pattern
-
-    def _create_dilated_sparse_pattern(self, num_blocks: int, device: torch.device) -> torch.Tensor:
-        """Create dilated sparsity pattern matching Ring Attention structure"""
-        pattern = torch.zeros(num_blocks, num_blocks, dtype=torch.bool, device=device)
-
-        # Multiple dilation rates for hierarchical attention
-        dilation_rates = [1, 2, 4, 8, 16]
-
-        for dilation in dilation_rates:
-            for i in range(num_blocks):
-                # Attend to positions at dilation intervals
-                for j in range(0, num_blocks, dilation):
-                    if abs(i - j) <= num_blocks * self.config.sparsity_ratio:
-                        pattern[i, j] = True
-
-        return pattern
-
-    def _create_global_local_pattern(self, num_blocks: int, device: torch.device) -> torch.Tensor:
-        """Create global + local attention pattern"""
-        pattern = torch.zeros(num_blocks, num_blocks, dtype=torch.bool, device=device)
-
-        global_blocks = self.config.global_tokens // self.config.block_size
-        local_blocks = self.config.local_window_size // self.config.block_size
-
-        # Global blocks attend to everything
-        pattern[:global_blocks, :] = True
-        pattern[:, :global_blocks] = True
-
-        # Local attention for remaining blocks
-        for i in range(global_blocks, num_blocks):
-            start = max(global_blocks, i - local_blocks // 2)
-            end = min(num_blocks, i + local_blocks // 2 + 1)
-            pattern[i, start:end] = True
-
-        return pattern
-
-    def _create_adaptive_pattern(self, num_blocks: int, device: torch.device) -> torch.Tensor:
-        """Create adaptive pattern based on attention history"""
-        # Start with dilated sparse pattern as base
-        pattern = self._create_dilated_sparse_pattern(num_blocks, device)
-
-        # Adapt based on previous attention patterns if available
-        if self.adaptive_history:
-            recent_attention = self.adaptive_history[-1]
-            if recent_attention.size(0) == num_blocks:
-                # Boost important regions from history
-                importance_scores = recent_attention.mean(dim=-1)  # Average across key blocks
-                threshold = torch.quantile(importance_scores, 1.0 - self.config.sparsity_ratio)
-                adaptive_boost = importance_scores > threshold
-
-                # Expand important query blocks
-                for i, is_important in enumerate(adaptive_boost):
-                    if is_important:
-                        pattern[i, :] = pattern[i, :] | (importance_scores > threshold * 0.5)
-
-        return pattern
-
-    def _enforce_target_sparsity(
-        self, pattern: torch.Tensor, target_sparsity: float
-    ) -> torch.Tensor:
-        """Ensure pattern meets target sparsity ratio"""
-        # Handle empty patterns
-        if pattern.numel() == 0:
-            return pattern
-
-        current_density = pattern.float().mean().item()
-
-        # Allow small tolerance (2%)
-        if abs(current_density - target_sparsity) < 0.02:
-            return pattern
-
-        if current_density > target_sparsity:
-            # Pattern too dense, remove some connections
-            num_remove = int((current_density - target_sparsity) * pattern.numel())
-
-            active_indices = torch.nonzero(pattern, as_tuple=False)
-            if len(active_indices) > num_remove:
-                # Randomly remove connections (preserving diagonal)
-                remove_indices = torch.randperm(len(active_indices))[:num_remove]
-                for idx in remove_indices:
-                    i, j = active_indices[idx]
-                    if i != j:  # Preserve diagonal
-                        pattern[i, j] = False
-
-        else:
-            # Pattern too sparse, add some connections
-            num_add = int((target_sparsity - current_density) * pattern.numel())
-            inactive_indices = torch.nonzero(~pattern, as_tuple=False)
-            if len(inactive_indices) > num_add:
-                # Add connections randomly
-                add_indices = torch.randperm(len(inactive_indices))[:num_add]
-                for idx in add_indices:
-                    i, j = inactive_indices[idx]
-                    pattern[i, j] = True
-
-        return pattern
-
-    def update_adaptive_history(self, attention_weights: torch.Tensor):
-        """Update adaptive pattern history with recent attention weights"""
-        if self.config.pattern_type == "adaptive":
-            # Store block-level attention summary
-            block_size = self.config.block_size
-            seq_len = attention_weights.size(-1)
-            num_blocks = seq_len // block_size
-
-            # Pool attention weights to block level
-            reshaped = attention_weights.view(-1, num_blocks, block_size, num_blocks, block_size)
-            block_attention = reshaped.mean(dim=(2, 4))  # Average within blocks
-
-            self.adaptive_history.append(block_attention.mean(dim=0))  # Average across batch/heads
-
-            # Keep only recent history
-            if len(self.adaptive_history) > 10:
-                self.adaptive_history.pop(0)
-
-
-class ContentAdaptiveSparsity(nn.Module):
-    """Neural network for learning content-adaptive sparsity patterns"""
-
-    def __init__(self, head_dim: int, block_size: int = 128, hidden_dim: int = None):
-        super().__init__()
-        self.head_dim = head_dim
-        self.block_size = block_size
-        self.hidden_dim = hidden_dim or head_dim // 4
-
-        # Importance predictor network
-        self.importance_predictor = nn.Sequential(
-            nn.Linear(head_dim, self.hidden_dim),
-            nn.ReLU(),
-            nn.Dropout(0.1),
-            nn.Linear(self.hidden_dim, self.hidden_dim // 2),
-            nn.ReLU(),
-            nn.Linear(self.hidden_dim // 2, 1),
-            nn.Sigmoid(),
-        )
-
-        # Block interaction predictor
-        self.interaction_predictor = nn.Sequential(
-            nn.Linear(head_dim * 2, self.hidden_dim),
-            nn.ReLU(),
-            nn.Linear(self.hidden_dim, 1),
-            nn.Sigmoid(),
-        )
-
-    def predict_block_importance(
-        self, q: torch.Tensor, k: torch.Tensor, sparsity_ratio: float = 0.25
-    ) -> torch.Tensor:
-        """Predict which block pairs are important for attention"""
-        batch, seq_len, heads, head_dim = q.shape
-        num_blocks = seq_len // self.block_size
-
-        # Reshape to blocks and average pool
-        q_blocks = q.view(batch, num_blocks, self.block_size, heads, head_dim)
-        k_blocks = k.view(batch, num_blocks, self.block_size, heads, head_dim)
-
-        q_block_avg = q_blocks.mean(dim=2)  # [batch, num_blocks, heads, head_dim]
-        k_block_avg = k_blocks.mean(dim=2)
-
-        # Predict individual block importance
-        q_importance = self.importance_predictor(q_block_avg)  # [batch, num_blocks, heads, 1]
-        k_importance = self.importance_predictor(k_block_avg)
-
-        # Predict block-pair interactions
-        q_expanded = q_block_avg.unsqueeze(2).expand(-1, -1, num_blocks, -1, -1)
-        k_expanded = k_block_avg.unsqueeze(1).expand(-1, num_blocks, -1, -1, -1)
-
-        # Concatenate Q and K for interaction prediction
-        qk_pairs = torch.cat([q_expanded, k_expanded], dim=-1)  # [..., head_dim * 2]
-        interaction_scores = self.interaction_predictor(qk_pairs).squeeze(
-            -1
-        )  # [batch, num_blocks, num_blocks, heads]
-
-        # Combine importance and interaction scores
-        combined_scores = (
-            q_importance.unsqueeze(2) * k_importance.unsqueeze(1) * interaction_scores.unsqueeze(-1)
-        ).squeeze(-1)
-
-        # Create sparse pattern by selecting top-k block pairs
-        k_pairs = int(num_blocks * num_blocks * sparsity_ratio)
-        flat_scores = combined_scores.view(batch, heads, -1)
-
-        patterns = torch.zeros_like(combined_scores, dtype=torch.bool)
-        for b in range(batch):
-            for h in range(heads):
-                _, top_indices = torch.topk(flat_scores[b, h], k_pairs)
-                # Convert flat indices to 2D coordinates
-                i_indices = top_indices // num_blocks
-                j_indices = top_indices % num_blocks
-                top_pairs = (i_indices, j_indices)
-                patterns[b, top_pairs[0], top_pairs[1], h] = True
-
-        return patterns.permute(0, 3, 1, 2)  # [batch, heads, num_blocks, num_blocks]
 
 
 class BlockSparseRingDilatedAttention(RingDilatedAttention):
     """
-    Block-Sparse Ring Dilated Attention with advanced optimization features.
+    Memory-efficient block-sparse ring dilated attention.
 
-    Combines the memory efficiency of Ring Attention with the computational
-    efficiency of block-sparse patterns for maximum performance.
-
-    Features:
-    - Multiple sparsity patterns (local window, dilated sparse, adaptive)
-    - 5-50x speedup over dense attention
-    - 75-95% memory reduction
-    - 95-99% quality retention
-    - Hardware-optimized kernels
-    - Dynamic pattern adaptation
+    Key improvements:
+    - Never materializes full attention matrices
+    - Returns sparse attention weights in COO format
+    - Processes only active blocks
+    - Minimal memory overhead
     """
 
     def __init__(
         self,
-        segment_lengths: Sequence[int],
-        dilation_rates: Sequence[int],
+        segment_lengths: list[int],
+        dilation_rates: list[int],
         sparse_config: SparsePatternConfig | None = None,
-        use_adaptive_sparsity: bool = False,
-        quality_threshold: float = 0.95,
-        enable_memory_pool: bool = True,
-        enable_packed_comm: bool = True,
-        enable_hardware_opt: bool = True,
         **kwargs,
     ):
-        """
-        Initialize Block-Sparse Ring Dilated Attention.
-
-        Args:
-            segment_lengths: Sequence of segment lengths for dilated attention
-            dilation_rates: Corresponding dilation rates
-            sparse_config: Configuration for sparsity patterns
-            use_adaptive_sparsity: Whether to use learned adaptive patterns
-            quality_threshold: Minimum quality threshold for adaptive sparsity
-            enable_memory_pool: Whether to use optimized memory pool
-            enable_packed_comm: Whether to use packed K/V communication
-            enable_hardware_opt: Whether to enable hardware-specific optimizations
-            **kwargs: Additional arguments passed to RingDilatedAttention
-        """
+        """Initialize block-sparse ring dilated attention."""
         # Extract sparse_config from kwargs if passed there (for compatibility)
         if "sparse_config" in kwargs:
             sparse_config = kwargs.pop("sparse_config")
         if "sparsity_config" in kwargs:
-            # Handle alternate name
             sparse_config = kwargs.pop("sparsity_config")
 
-        # Filter out BlockSparse-specific parameters before passing to parent
+        # Filter out any remaining BlockSparse-specific parameters
         block_sparse_params = {
             "use_adaptive_sparsity",
             "quality_threshold",
             "enable_memory_pool",
             "enable_packed_comm",
             "enable_hardware_opt",
-            "sparse_config",
-            "sparsity_config",  # Make sure these don't get passed to parent
         }
         filtered_kwargs = {k: v for k, v in kwargs.items() if k not in block_sparse_params}
 
         super().__init__(segment_lengths, dilation_rates, **filtered_kwargs)
 
-        # Validate quality threshold
-        if not 0.0 <= quality_threshold <= 1.0:
-            raise ValueError(f"quality_threshold must be between 0 and 1, got {quality_threshold}")
-
-        # Sparsity configuration
         self.sparse_config = sparse_config or SparsePatternConfig()
+        self.block_size = self.sparse_config.block_size
 
-        # Validate sparse config
-        if self.sparse_config.sparsity_ratio <= 0.0 or self.sparse_config.sparsity_ratio >= 1.0:
-            raise ValueError(
-                f"sparsity_ratio must be between 0 and 1 (exclusive), "
-                f"got {self.sparse_config.sparsity_ratio}"
-            )
-
-        if self.sparse_config.block_size <= 0:
-            raise ValueError(f"block_size must be positive, got {self.sparse_config.block_size}")
-
-        if self.sparse_config.local_window_size <= 0:
-            raise ValueError(
-                f"local_window_size must be positive, got {self.sparse_config.local_window_size}"
-            )
-
-        self.use_adaptive_sparsity = use_adaptive_sparsity
-        self.quality_threshold = quality_threshold
-
-        # Optimization flags
-        self.enable_memory_pool = enable_memory_pool
-        self.enable_packed_comm = enable_packed_comm
-        self.enable_hardware_opt = enable_hardware_opt
-
-        # Pattern generator
-        self.pattern_generator = SparsePatternGenerator(self.sparse_config)
-
-        # Memory pool
-        if self.enable_memory_pool:
-            self.memory_pool = BlockSparseMemoryPool(max_pool_size=50, hot_cache_size=10)
-        else:
-            self.memory_pool = None
-
-        # Hardware detection
-        if self.enable_hardware_opt:
-            self._detect_hardware()
-        else:
-            self.is_h100 = False
-            self.has_fa3 = False
-
-        # Adaptive sparsity network (if enabled)
-        self.adaptive_sparsity = None
-        if self.use_adaptive_sparsity:
-            # Initialize with reasonable head_dim, will be updated on first forward pass
-            self.adaptive_sparsity = ContentAdaptiveSparsity(
-                head_dim=64,  # Will be updated dynamically
-                block_size=self.sparse_config.block_size,
-            )
-
-        # Performance tracking
-        self.performance_stats = {
-            "total_forwards": 0,
-            "sparse_ratio_history": [],
-            "quality_scores": [],
-            "speedup_ratios": [],
-        }
-
-        # Thread-safe locks
-        self._stats_lock = threading.Lock()
-        self._buffer_lock = threading.Lock()
-        self._pattern_lock = threading.Lock()
-
-        # Communication optimization
-        if self.enable_packed_comm and hasattr(self, "ring_size") and self.ring_size > 1:
-            self._init_packed_comm_buffers()
-
-        # Versioned pattern cache
-        self.pattern_cache_version = 0
-        self.max_pattern_cache_size = 25
-
-    def _detect_hardware(self):
-        """Detect hardware capabilities for optimization."""
-        if torch.cuda.is_available():
-            device_name = torch.cuda.get_device_name()
-            self.is_h100 = "H100" in device_name
-
-            # Check Flash Attention 3 support
-            self.has_fa3 = False
-            if HAS_FLASH_ATTN:
-                try:
-                    # FA3 supports H100 optimizations
-                    import flash_attn
-
-                    if hasattr(flash_attn, "__version__"):
-                        version = flash_attn.__version__
-                        major = int(version.split(".")[0])
-                        self.has_fa3 = major >= 3 and self.is_h100
-                except Exception:
-                    pass
-        else:
-            self.is_h100 = False
-            self.has_fa3 = False
-
-    def _init_packed_comm_buffers(self):
-        """Initialize buffers for packed K/V communication."""
-        # Pre-allocate communication buffers
-        self.send_buffer = None
-        self.recv_buffer = None
-        self.pack_stream = None
-        self.comm_stream = None
-
-        if torch.cuda.is_available() and hasattr(self, "use_async_comm") and self.use_async_comm:
-            self.pack_stream = torch.cuda.Stream()
-            self.comm_stream = torch.cuda.Stream()
-
-    def _get_buffer(
-        self, shape: tuple[int, ...], dtype: torch.dtype, device: torch.device
-    ) -> Tensor:
-        """Get buffer with memory pool support."""
-        if self.memory_pool is not None:
-            return self.memory_pool.get_buffer(shape, dtype, device)
-        else:
-            return torch.empty(shape, dtype=dtype, device=device)
-
-    def _return_buffer(self, buffer: Tensor):
-        """Return buffer to memory pool."""
-        if self.memory_pool is not None:
-            self.memory_pool.return_buffer(buffer)
+        # Pattern cache - stores only block indices, not full matrices
+        self.pattern_cache: dict[tuple, tuple[Tensor, Tensor]] = {}
+        self._cache_lock = threading.Lock()
 
     def forward(
         self,
@@ -685,506 +80,288 @@ class BlockSparseRingDilatedAttention(RingDilatedAttention):
         v: Tensor,
         is_causal: bool = False,
         return_attention_weights: bool = False,
-    ) -> Tensor | tuple[Tensor, Tensor]:
+    ) -> Tensor | tuple[Tensor, dict[str, Tensor] | None]:
         """
-        Forward pass with block-sparse ring attention.
-
-        Args:
-            q: Query tensor [batch, seq_len, num_heads, head_dim]
-            k: Key tensor [batch, seq_len, num_heads, head_dim]
-            v: Value tensor [batch, seq_len, num_heads, head_dim]
-            is_causal: Whether to apply causal masking
-            return_attention_weights: Whether to return attention weights
+        Forward pass with memory-efficient block-sparse attention.
 
         Returns:
             output: Attention output [batch, seq_len, num_heads, head_dim]
-            attention_weights: Optional attention weights (if requested)
+            attention_weights: If requested, returns dict with:
+                - 'indices': COO format indices of non-zero blocks
+                - 'values': Attention values for those blocks
+                - 'shape': Full attention shape for reference
         """
         batch, seq_len, num_heads, head_dim = q.shape
 
-        # Update adaptive sparsity network dimensions if needed
-        if self.use_adaptive_sparsity and self.adaptive_sparsity.head_dim != head_dim:
-            self.adaptive_sparsity = ContentAdaptiveSparsity(
-                head_dim=head_dim, block_size=self.sparse_config.block_size
-            ).to(q.device)
+        # Ensure sequence length is divisible by block size
+        if seq_len % self.block_size != 0:
+            raise ValueError(
+                f"Sequence length {seq_len} must be divisible by block size {self.block_size}"
+            )
 
-        # Create or retrieve sparse pattern
-        sparse_pattern = self._get_sparse_pattern(q, k)
+        num_blocks = seq_len // self.block_size
 
-        # Apply block-sparse ring attention
-        output, attention_weights = self._block_sparse_ring_attention(
-            q, k, v, sparse_pattern, is_causal, return_attention_weights
+        # Get active block pairs for sparse pattern
+        block_indices = self._get_sparse_block_indices(num_blocks, num_heads, q.device)
+
+        # Initialize output
+        output = torch.zeros_like(q)
+
+        # Process attention in blocks
+        if return_attention_weights:
+            attention_data = self._compute_sparse_attention_with_weights(
+                q, k, v, output, block_indices, is_causal
+            )
+        else:
+            self._compute_sparse_attention(q, k, v, output, block_indices, is_causal)
+            attention_data = None
+
+        return (output, attention_data) if return_attention_weights else output
+
+    def _get_sparse_block_indices(
+        self, num_blocks: int, num_heads: int, device: torch.device
+    ) -> tuple[Tensor, Tensor]:
+        """Get indices of active blocks in sparse pattern."""
+        _ = num_heads  # Reserved for future head-specific patterns
+        cache_key = (
+            num_blocks,
+            self.sparse_config.pattern_type,
+            self.sparse_config.sparsity_ratio,
+            self.sparse_config.block_size,
         )
 
-        # Update performance statistics
-        self._update_performance_stats(sparse_pattern, attention_weights)
+        with self._cache_lock:
+            if cache_key in self.pattern_cache:
+                row_idx, col_idx = self.pattern_cache[cache_key]
+                return row_idx.to(device), col_idx.to(device)
 
-        # Update adaptive pattern history if applicable
-        if attention_weights is not None and self.sparse_config.pattern_type == "adaptive":
-            self.pattern_generator.update_adaptive_history(attention_weights)
-
-        if return_attention_weights:
-            return output, attention_weights
-        return output
-
-    def _get_sparse_pattern(self, q: Tensor, k: Tensor) -> Tensor:
-        """Get or create sparse attention pattern"""
-        seq_len, num_heads = q.size(1), q.size(2)
-
-        if self.use_adaptive_sparsity and self.adaptive_sparsity is not None:
-            # Use learned adaptive pattern
-            pattern = self.adaptive_sparsity.predict_block_importance(
-                q, k, self.sparse_config.sparsity_ratio
-            )
-            return pattern
+        # Generate pattern based on type
+        if self.sparse_config.pattern_type == "local_window":
+            row_idx, col_idx = self._create_local_window_indices(num_blocks)
+        elif self.sparse_config.pattern_type == "dilated_sparse":
+            row_idx, col_idx = self._create_dilated_sparse_indices(num_blocks)
+        elif self.sparse_config.pattern_type == "global_local":
+            row_idx, col_idx = self._create_global_local_indices(num_blocks)
         else:
-            # Use predefined pattern
-            pattern = self.pattern_generator.create_pattern(seq_len, num_heads, q.device)
-            # Expand for batch and heads if needed
-            if pattern.dim() == 2:
-                pattern = pattern.unsqueeze(0).unsqueeze(0).expand(q.size(0), num_heads, -1, -1)
-            return pattern
+            raise ValueError(f"Unknown pattern type: {self.sparse_config.pattern_type}")
 
-    def _block_sparse_ring_attention(
+        # Convert to tensors
+        row_idx = torch.tensor(row_idx, dtype=torch.long)
+        col_idx = torch.tensor(col_idx, dtype=torch.long)
+
+        # Cache the indices (on CPU to save GPU memory)
+        with self._cache_lock:
+            self.pattern_cache[cache_key] = (row_idx.cpu(), col_idx.cpu())
+
+        return row_idx.to(device), col_idx.to(device)
+
+    def _create_local_window_indices(self, num_blocks: int) -> tuple[list, list]:
+        """Create indices for local window pattern."""
+        window_blocks = self.sparse_config.local_window_size // self.block_size
+        row_indices = []
+        col_indices = []
+
+        for i in range(num_blocks):
+            start = max(0, i - window_blocks // 2)
+            end = min(num_blocks, i + window_blocks // 2 + 1)
+            for j in range(start, end):
+                row_indices.append(i)
+                col_indices.append(j)
+
+        return row_indices, col_indices
+
+    def _create_dilated_sparse_indices(self, num_blocks: int) -> tuple[list, list]:
+        """Create indices for dilated sparse pattern."""
+        row_indices = []
+        col_indices = []
+
+        # Use dilation rates from parent class
+        for i in range(num_blocks):
+            # Always attend to self
+            row_indices.append(i)
+            col_indices.append(i)
+
+            # Add dilated connections
+            for rate in self.dilation_rates[:3]:  # Use first 3 dilation rates
+                for direction in [-1, 1]:
+                    j = i + direction * rate
+                    if 0 <= j < num_blocks:
+                        row_indices.append(i)
+                        col_indices.append(j)
+
+        # Remove duplicates while preserving order
+        seen = set()
+        unique_pairs = []
+        for r, c in zip(row_indices, col_indices, strict=False):
+            if (r, c) not in seen:
+                seen.add((r, c))
+                unique_pairs.append((r, c))
+
+        if unique_pairs:
+            row_indices, col_indices = zip(*unique_pairs, strict=False)
+        else:
+            row_indices, col_indices = [], []
+
+        return list(row_indices), list(col_indices)
+
+    def _create_global_local_indices(self, num_blocks: int) -> tuple[list, list]:
+        """Create indices for global + local pattern."""
+        row_indices = []
+        col_indices = []
+
+        global_blocks = min(self.sparse_config.global_tokens // self.block_size, num_blocks)
+        local_radius = (self.sparse_config.local_window_size // self.block_size) // 2
+
+        for i in range(num_blocks):
+            # Global attention to first few blocks
+            for j in range(global_blocks):
+                row_indices.append(i)
+                col_indices.append(j)
+
+            # Local window attention
+            start = max(global_blocks, i - local_radius)
+            end = min(num_blocks, i + local_radius + 1)
+            for j in range(start, end):
+                if j not in range(global_blocks):  # Avoid duplicates
+                    row_indices.append(i)
+                    col_indices.append(j)
+
+        return row_indices, col_indices
+
+    def _compute_sparse_attention(
         self,
         q: Tensor,
         k: Tensor,
         v: Tensor,
-        sparse_pattern: Tensor,
+        output: Tensor,
+        block_indices: tuple[Tensor, Tensor],
         is_causal: bool,
-        return_attention_weights: bool,
-    ) -> tuple[Tensor, Tensor | None]:
-        """Execute block-sparse ring attention computation"""
+    ):
+        """Compute sparse attention without storing weights."""
         batch, seq_len, num_heads, head_dim = q.shape
-        block_size = self.sparse_config.block_size
+        row_idx, col_idx = block_indices
+        num_active_blocks = len(row_idx)
 
-        # Check for Flash Attention 3 optimization
-        if self.has_fa3 and HAS_FLASH_ATTN and not return_attention_weights:
-            return self._fa3_block_sparse_attention(q, k, v, sparse_pattern, is_causal)
+        if num_active_blocks == 0:
+            return
 
-        # Calculate number of blocks
-        num_blocks = seq_len // block_size
+        # Reshape to blocks
+        q_blocks = q.view(batch, -1, self.block_size, num_heads, head_dim)
+        k_blocks = k.view(batch, -1, self.block_size, num_heads, head_dim)
+        v_blocks = v.view(batch, -1, self.block_size, num_heads, head_dim)
+        output_blocks = output.view(batch, -1, self.block_size, num_heads, head_dim)
 
-        # Initialize output tensor with memory pool
-        output = self._get_buffer(q.shape, q.dtype, q.device)
-        output.zero_()
-        attention_weights_full = None
-
-        if return_attention_weights:
-            attention_weights_full = self._get_buffer(
-                (batch, num_heads, seq_len, seq_len), q.dtype, q.device
-            )
-            attention_weights_full.zero_()
-
-        # Reshape inputs to blocks
-        q_blocks = q.view(batch, num_blocks, block_size, num_heads, head_dim)
-        k_blocks = k.view(batch, num_blocks, block_size, num_heads, head_dim)
-        v_blocks = v.view(batch, num_blocks, block_size, num_heads, head_dim)
-
-        # Process each ring step
-        for ring_step in range(self.ring_size):
-            # Get K/V for current ring step
-            ring_k, ring_v = self._get_ring_kv_blocks(k_blocks, v_blocks, ring_step)
-
-            # Process sparse blocks for this ring step
-            step_output, step_weights = self._process_sparse_ring_step(
-                q_blocks,
-                ring_k,
-                ring_v,
-                sparse_pattern,
-                ring_step,
-                is_causal,
-                return_attention_weights,
-            )
-
-            # Accumulate output
-            output += step_output.view(batch, seq_len, num_heads, head_dim)
-
-            if return_attention_weights and step_weights is not None:
-                attention_weights_full += step_weights
-
-        # Normalize by ring size
-        output = output / self.ring_size
-        if return_attention_weights and attention_weights_full is not None:
-            attention_weights_full = attention_weights_full / self.ring_size
-
-        # Clone output before returning buffers
-        output_final = output.clone()
-        self._return_buffer(output)
-
-        if return_attention_weights and attention_weights_full is not None:
-            weights_final = attention_weights_full.clone()
-            self._return_buffer(attention_weights_full)
-            return output_final, weights_final
-        else:
-            # Make sure to return buffer even if not returning weights
-            if attention_weights_full is not None:
-                self._return_buffer(attention_weights_full)
-            return output_final, None
-
-    def _fa3_block_sparse_attention(
-        self, q: Tensor, k: Tensor, v: Tensor, sparse_pattern: Tensor, is_causal: bool
-    ) -> tuple[Tensor, None]:
-        """Flash Attention 3 optimized block-sparse attention."""
-        # Convert block sparse pattern to FA3 format
-        batch, seq_len, num_heads, head_dim = q.shape
-        block_size = self.sparse_config.block_size
-
-        # Validate FA3 compatibility
-        if self.sparse_config.pattern_type not in ["local_window", "dilated_sparse"]:
-            warnings.warn(
-                f"Flash Attention 3 may not fully support pattern type '{self.sparse_config.pattern_type}'. "
-                "Falling back to standard implementation."
-            )
-            # Fall back to standard sparse attention
-            return self._block_sparse_ring_attention(q, k, v, sparse_pattern, is_causal, False)
-
-        # Validate block size for FA3
-        if block_size % 64 != 0 and self.is_h100:
-            warnings.warn(
-                f"Block size {block_size} is not optimal for H100. "
-                "Consider using multiples of 64 for better performance."
-            )
-
-        # Reshape for FA3
-        q_fa3 = q.transpose(1, 2).contiguous()  # [batch, num_heads, seq_len, head_dim]
-        k_fa3 = k.transpose(1, 2).contiguous()
-        v_fa3 = v.transpose(1, 2).contiguous()
-
-        try:
-            # Use FA3 with appropriate configuration
-            if self.sparse_config.pattern_type == "local_window":
-                # Local window attention
-                output_fa3 = flash_attn_func(
-                    q_fa3,
-                    k_fa3,
-                    v_fa3,
-                    causal=is_causal,
-                    window_size=(block_size, block_size),
-                )
-            elif self.sparse_config.pattern_type == "dilated_sparse":
-                # For dilated sparse, FA3 doesn't have direct support
-                # Use standard FA3 without window for now
-                output_fa3 = flash_attn_func(q_fa3, k_fa3, v_fa3, causal=is_causal)
-            else:
-                # Fallback
-                raise RuntimeError("Unsupported pattern type for FA3")
-
-        except Exception as e:
-            warnings.warn(
-                f"Flash Attention 3 failed with error: {e}. Falling back to standard implementation."
-            )
-            # Fall back to standard implementation
-            return self._block_sparse_ring_attention(q, k, v, sparse_pattern, is_causal, False)
-
-        # Reshape back
-        output = output_fa3.transpose(1, 2).contiguous()  # [batch, seq_len, num_heads, head_dim]
-        return output, None
-
-    def _process_sparse_ring_step(
-        self,
-        q_blocks: Tensor,
-        k_blocks: Tensor,
-        v_blocks: Tensor,
-        sparse_pattern: Tensor,
-        ring_step: int,
-        is_causal: bool,
-        return_attention_weights: bool,
-    ) -> tuple[Tensor, Tensor | None]:
-        """Process sparse blocks for a single ring step - OPTIMIZED VERSION"""
-        batch, num_blocks, block_size, num_heads, head_dim = q_blocks.shape
-
-        # Initialize output
-        output_blocks = torch.zeros_like(q_blocks)
-        attention_weights = None
-
-        if return_attention_weights:
-            seq_len = num_blocks * block_size
-            attention_weights = torch.zeros(
-                batch,
-                num_heads,
-                seq_len,
-                seq_len,
-                device=q_blocks.device,
-                dtype=q_blocks.dtype,
-            )
-
-        # Find active block pairs for this ring step
-        ring_pattern = self._get_ring_step_pattern(sparse_pattern, ring_step)
-
-        # OPTIMIZATION: Process all blocks for each head in parallel
-        # This avoids the expensive Python loop over individual blocks
+        # Process each active block pair
         scale = 1.0 / math.sqrt(head_dim)
 
-        for head_idx in range(num_heads):
-            # Get pattern for this head
-            if ring_pattern.dim() == 4:  # [batch, heads, blocks, blocks]
-                head_pattern = ring_pattern[:, head_idx]
-            else:  # Assume already head-specific
-                head_pattern = ring_pattern
+        for idx in range(num_active_blocks):
+            q_idx = row_idx[idx]
+            k_idx = col_idx[idx]
 
-            # Find all active block pairs for this head
-            active_indices = head_pattern.nonzero(as_tuple=True)
+            # Extract blocks
+            q_block = q_blocks[:, q_idx]  # [batch, block_size, num_heads, head_dim]
+            k_block = k_blocks[:, k_idx]
+            v_block = v_blocks[:, k_idx]
 
-            if len(active_indices[0]) == 0:
-                continue  # No active blocks
+            # Compute attention for this block
+            # Reshape for batch matrix multiply
+            q_block = q_block.transpose(1, 2)  # [batch, num_heads, block_size, head_dim]
+            k_block = k_block.transpose(1, 2)
+            v_block = v_block.transpose(1, 2)
 
-            batch_indices, q_block_indices, k_block_indices = active_indices
-            num_active = len(batch_indices)
-
-            # Extract all active blocks at once - much more efficient
-            q_active = q_blocks[batch_indices, q_block_indices, :, head_idx, :]
-            k_active = k_blocks[batch_indices, k_block_indices, :, head_idx, :]
-            v_active = v_blocks[batch_indices, k_block_indices, :, head_idx, :]
-
-            # Batched attention computation
-            scores = torch.bmm(q_active, k_active.transpose(-2, -1)) * scale
+            # Attention scores
+            scores = torch.matmul(q_block, k_block.transpose(-2, -1)) * scale
 
             # Apply causal mask if needed
             if is_causal:
-                causal_mask = torch.triu(
-                    torch.ones(block_size, block_size, device=scores.device, dtype=torch.bool),
-                    diagonal=1,
-                )
-                scores.masked_fill_(causal_mask, float("-inf"))
+                # Only apply mask if this is a causal block (q_idx >= k_idx)
+                if q_idx > k_idx:
+                    # Full block is allowed
+                    pass
+                elif q_idx == k_idx:
+                    # Diagonal block needs causal mask
+                    causal_mask = torch.triu(
+                        torch.ones(self.block_size, self.block_size, device=scores.device),
+                        diagonal=1,
+                    ).bool()
+                    scores.masked_fill_(causal_mask, float("-inf"))
+                else:
+                    # Skip future blocks entirely
+                    continue
+
+            # Softmax and attention
+            attn_weights = torch.softmax(scores, dim=-1)
+            attn_output = torch.matmul(attn_weights, v_block)
+
+            # Accumulate to output (back to original shape)
+            attn_output = attn_output.transpose(1, 2)  # [batch, block_size, num_heads, head_dim]
+            output_blocks[:, q_idx] += attn_output
+
+    def _compute_sparse_attention_with_weights(
+        self,
+        q: Tensor,
+        k: Tensor,
+        v: Tensor,
+        output: Tensor,
+        block_indices: tuple[Tensor, Tensor],
+        is_causal: bool,
+    ) -> dict[str, Tensor]:
+        """Compute sparse attention and return weights in sparse format."""
+        batch, seq_len, num_heads, head_dim = q.shape
+        row_idx, col_idx = block_indices
+
+        # Prepare sparse storage for attention weights
+        # We'll store block indices and their attention values
+        weight_indices = []
+        weight_values = []
+
+        # Reshape to blocks
+        q_blocks = q.view(batch, -1, self.block_size, num_heads, head_dim)
+        k_blocks = k.view(batch, -1, self.block_size, num_heads, head_dim)
+        v_blocks = v.view(batch, -1, self.block_size, num_heads, head_dim)
+        output_blocks = output.view(batch, -1, self.block_size, num_heads, head_dim)
+
+        scale = 1.0 / math.sqrt(head_dim)
+
+        for idx in range(len(row_idx)):
+            q_idx = row_idx[idx].item()
+            k_idx = col_idx[idx].item()
+
+            # Skip future blocks for causal attention
+            if is_causal and q_idx < k_idx:
+                continue
+
+            # Extract blocks
+            q_block = q_blocks[:, q_idx].transpose(1, 2)
+            k_block = k_blocks[:, k_idx].transpose(1, 2)
+            v_block = v_blocks[:, k_idx].transpose(1, 2)
 
             # Compute attention
-            attn_probs = F.softmax(scores, dim=-1)
-            block_outputs = torch.bmm(attn_probs, v_active)
+            scores = torch.matmul(q_block, k_block.transpose(-2, -1)) * scale
 
-            # Accumulate results
-            output_blocks[batch_indices, q_block_indices, :, head_idx, :] += block_outputs
+            # Apply causal mask if needed
+            if is_causal and q_idx == k_idx:
+                causal_mask = torch.triu(
+                    torch.ones(self.block_size, self.block_size, device=scores.device), diagonal=1
+                ).bool()
+                scores.masked_fill_(causal_mask, float("-inf"))
 
-            # Store attention weights if requested
-            if return_attention_weights:
-                for idx in range(num_active):
-                    b_idx = batch_indices[idx]
-                    q_idx = q_block_indices[idx]
-                    k_idx = k_block_indices[idx]
-                    q_start = q_idx * block_size
-                    q_end = (q_idx + 1) * block_size
-                    k_start = k_idx * block_size
-                    k_end = (k_idx + 1) * block_size
-                    attention_weights[b_idx, head_idx, q_start:q_end, k_start:k_end] = attn_probs[
-                        idx
-                    ]
+            # Softmax
+            attn_weights = torch.softmax(scores, dim=-1)
 
-        return output_blocks, attention_weights
+            # Store sparse weights (only the block, not full matrix)
+            weight_indices.append((q_idx, k_idx))
+            weight_values.append(attn_weights.cpu())  # Store on CPU to save GPU memory
 
-    def _compute_block_attention(
-        self,
-        q_block: Tensor,
-        k_block: Tensor,
-        v_block: Tensor,
-        is_causal: bool,
-        return_weights: bool,
-    ) -> tuple[Tensor, Tensor | None]:
-        """Compute attention for a single block pair"""
-        scale = 1.0 / math.sqrt(q_block.size(-1))
+            # Compute output
+            attn_output = torch.matmul(attn_weights, v_block)
+            output_blocks[:, q_idx] += attn_output.transpose(1, 2)
 
-        # Compute attention scores
-        scores = torch.matmul(q_block, k_block.transpose(-2, -1)) * scale
-
-        # Apply causal mask if needed
-        if is_causal:
-            block_size = q_block.size(0)
-            causal_mask = torch.triu(
-                torch.ones(block_size, block_size, device=q_block.device), diagonal=1
-            )
-            scores = scores.masked_fill(causal_mask.bool(), float("-inf"))
-
-        # Apply softmax
-        attention_weights = torch.softmax(scores, dim=-1)
-
-        # Compute output
-        output = torch.matmul(attention_weights, v_block)
-
-        if return_weights:
-            return output, attention_weights
-        return output, None
-
-    def _get_ring_kv_blocks(
-        self, k_blocks: Tensor, v_blocks: Tensor, ring_step: int
-    ) -> tuple[Tensor, Tensor]:
-        """Get K/V blocks for current ring step with sparse communication"""
-        if self.ring_size == 1:
-            return k_blocks, v_blocks
-
-        # Rotate blocks according to ring step
-        rotation = ring_step % self.ring_size
-        if rotation == 0:
-            return k_blocks, v_blocks
-
-        # Check if distributed is properly initialized
-        if torch.distributed.is_initialized():
-            # Use packed communication if enabled
-            if self.enable_packed_comm:
-                return self._ring_rotate_kv_packed(k_blocks, v_blocks)
-            else:
-                # Use standard distributed communication
-                warnings.warn("Ring attention without packed communication may be slower")
-                return self._ring_rotate_kv_standard(k_blocks, v_blocks, rotation)
-        else:
-            # Single GPU case - ring_size should be 1
-            if self.ring_size > 1:
-                raise RuntimeError(
-                    f"Ring size is {self.ring_size} but distributed is not initialized. "
-                    "Either initialize distributed training or set ring_size=1."
-                )
-            # This should not be reached due to ring_size check above
-            return k_blocks, v_blocks
-
-    def _ring_rotate_kv_packed(self, k: Tensor, v: Tensor) -> tuple[Tensor, Tensor]:
-        """Rotate K/V blocks with packed communication."""
-        if not torch.distributed.is_initialized():
-            return k, v
-
-        with self._buffer_lock:
-            # Pack K and V
-            pack_shape = (k.shape[0], k.shape[1], 2, k.shape[2], k.shape[3], k.shape[4])
-            if self.send_buffer is None or self.send_buffer.shape != pack_shape:
-                if self.send_buffer is not None:
-                    self._return_buffer(self.send_buffer)
-                self.send_buffer = self._get_buffer(pack_shape, k.dtype, k.device)
-
-            # Pack K and V efficiently
-            self.send_buffer[:, :, 0] = k
-            self.send_buffer[:, :, 1] = v
-
-            # Get receive buffer
-            if self.recv_buffer is None or self.recv_buffer.shape != pack_shape:
-                if self.recv_buffer is not None:
-                    self._return_buffer(self.recv_buffer)
-                self.recv_buffer = self._get_buffer(pack_shape, k.dtype, k.device)
-
-            # Ring communication
-            if hasattr(self, "rank"):
-                dist.send(self.send_buffer, dst=(self.rank + 1) % self.ring_size)
-                dist.recv(self.recv_buffer, src=(self.rank - 1) % self.ring_size)
-
-            # Unpack
-            k_new = self.recv_buffer[:, :, 0].contiguous()
-            v_new = self.recv_buffer[:, :, 1].contiguous()
-
-            return k_new, v_new
-
-    def _ring_rotate_kv_standard(
-        self, k: Tensor, v: Tensor, rotation: int
-    ) -> tuple[Tensor, Tensor]:
-        """Standard distributed rotation without packing."""
-        # Allocate temporary buffers
-        k_send = k.contiguous()
-        v_send = v.contiguous()
-        k_recv = torch.empty_like(k)
-        v_recv = torch.empty_like(v)
-
-        # Send/receive K and V separately
-        if hasattr(self, "rank"):
-            send_rank = (self.rank + 1) % self.ring_size
-            recv_rank = (self.rank - 1) % self.ring_size
-
-            # Use non-blocking operations for better performance
-            k_send_op = dist.isend(k_send, dst=send_rank)
-            k_recv_op = dist.irecv(k_recv, src=recv_rank)
-            v_send_op = dist.isend(v_send, dst=send_rank)
-            v_recv_op = dist.irecv(v_recv, src=recv_rank)
-
-            # Wait for all operations to complete
-            k_send_op.wait()
-            k_recv_op.wait()
-            v_send_op.wait()
-            v_recv_op.wait()
-
-            return k_recv, v_recv
-        else:
-            # Fallback if rank not set
-            warnings.warn("Rank not set for ring communication, returning original tensors")
-            return k, v
-
-    def _get_ring_step_pattern(self, sparse_pattern: Tensor, ring_step: int) -> Tensor:
-        """Get sparse pattern for specific ring step"""
-        # For simplicity, use the same pattern for all ring steps
-        # In practice, could rotate or adapt pattern based on ring step
-        return sparse_pattern
-
-    def _update_performance_stats(self, sparse_pattern: Tensor, attention_weights: Tensor | None):
-        """Update performance tracking statistics"""
-        with self._stats_lock:
-            self.performance_stats["total_forwards"] += 1
-
-            # Calculate actual sparsity ratio
-            actual_sparsity = sparse_pattern.float().mean().item()
-            self.performance_stats["sparse_ratio_history"].append(actual_sparsity)
-
-            # Estimate speedup (theoretical based on sparsity)
-            # Validate sparsity is in valid range [0, 1]
-            if 0.0 <= actual_sparsity <= 1.0:
-                # Clamp to minimum reasonable sparsity to avoid extreme speedup values
-                safe_sparsity = max(actual_sparsity, 0.01)
-                speedup_ratio = 1.0 / safe_sparsity
-            else:
-                # Invalid sparsity, log warning and use default
-                warnings.warn(f"Invalid sparsity ratio {actual_sparsity}, using 1.0")
-                speedup_ratio = 1.0
-
-            self.performance_stats["speedup_ratios"].append(speedup_ratio)
-
-            # Keep only recent history
-            max_history = 100
-            for key in ["sparse_ratio_history", "speedup_ratios"]:
-                if len(self.performance_stats[key]) > max_history:
-                    self.performance_stats[key] = self.performance_stats[key][-max_history:]
-
-    def get_performance_stats(self) -> dict[str, Any]:
-        """Get current performance statistics"""
-        with self._stats_lock:
-            stats = self.performance_stats.copy()
-
-            if stats["sparse_ratio_history"]:
-                stats["avg_sparsity"] = sum(stats["sparse_ratio_history"]) / len(
-                    stats["sparse_ratio_history"]
-                )
-                stats["avg_speedup"] = sum(stats["speedup_ratios"]) / len(stats["speedup_ratios"])
-            else:
-                stats["avg_sparsity"] = 0.0
-                stats["avg_speedup"] = 1.0
-
-            return stats
-
-    def set_sparsity_ratio(self, sparsity_ratio: float):
-        """Dynamically adjust sparsity ratio"""
-        if not (0.01 <= sparsity_ratio <= 0.99):
-            raise ValueError(f"Sparsity ratio must be between 0.01 and 0.99, got {sparsity_ratio}")
-
-        self.sparse_config.sparsity_ratio = sparsity_ratio
-        # Clear pattern cache to force regeneration
-        self.pattern_generator.pattern_cache.clear()
-
-    def enable_adaptive_sparsity(self, enable: bool = True):
-        """Enable or disable adaptive sparsity learning"""
-        self.use_adaptive_sparsity = enable
-        if enable and self.adaptive_sparsity is None:
-            # Initialize adaptive sparsity network
-            self.adaptive_sparsity = ContentAdaptiveSparsity(
-                head_dim=64,  # Will be updated on first forward pass
-                block_size=self.sparse_config.block_size,
-            )
-
-    def get_memory_info(self) -> dict[str, Any]:
-        """Get comprehensive memory and optimization information"""
-        base_info = super().get_memory_info()
-
-        # Add sparse attention specific information
-        sparse_info = {
-            "sparse_pattern_type": self.sparse_config.pattern_type,
-            "sparsity_ratio": self.sparse_config.sparsity_ratio,
-            "block_size": self.sparse_config.block_size,
-            "theoretical_speedup": f"{1.0 / self.sparse_config.sparsity_ratio:.1f}x",
-            "memory_reduction": f"{(1.0 - self.sparse_config.sparsity_ratio) * 100:.1f}%",
-            "adaptive_sparsity_enabled": self.use_adaptive_sparsity,
-            "pattern_cache_size": len(self.pattern_generator.pattern_cache),
+        # Return sparse representation
+        return {
+            "block_indices": weight_indices,
+            "block_values": weight_values,
+            "block_size": self.block_size,
+            "shape": (batch, num_heads, seq_len, seq_len),
+            "num_blocks": (seq_len // self.block_size, seq_len // self.block_size),
         }
-
-        # Merge with base information
-        base_info.update(sparse_info)
-        return base_info
