+++ dilated_attention_pytorch/block_sparse_ring_multihead_dilated_attention.py
@@ -1,572 +1,175 @@
 """
-Block-Sparse Ring Multihead Dilated Attention Implementation
+Block-Sparse Ring Multihead Dilated Attention - Memory Efficient Implementation
 
-This module provides a drop-in replacement for nn.MultiheadAttention with
-block-sparse optimization and Ring Attention scaling. Designed for production
-use with maximum compatibility and performance.
-
-Key Features:
-- Drop-in replacement for PyTorch MultiheadAttention
-- Fused QKV projections with sparse attention patterns
-- Advanced memory management and buffer reuse
-- Hardware-optimized execution paths
-- Comprehensive monitoring and debugging tools
-- Production-ready error handling and recovery
-
-Performance Benefits:
-- 10-100x speedup over standard MultiheadAttention
-- 90-99% memory reduction for long sequences
-- Near-perfect quality retention (98-99%)
-- Linear scaling to unlimited sequence lengths
+Multihead wrapper for the memory-efficient block-sparse implementation.
 """
 
-import threading
-import warnings
-from collections.abc import Sequence
 from typing import Any
 
-import torch
 from torch import Tensor, nn
 
-# Import base implementations
 from .block_sparse_ring_dilated_attention import (
     BlockSparseRingDilatedAttention,
     SparsePatternConfig,
 )
 
 
-class FusedQKVProjection(nn.Module):
-    """
-    Fused QKV projection optimized for block-sparse attention.
-
-    Combines Q, K, V projections into a single operation with optimized
-    memory layout for sparse attention computation.
-    """
-
-    def __init__(
-        self,
-        embed_dim: int,
-        num_heads: int,
-        bias: bool = True,
-        dropout: float = 0.0,
-        device: torch.device | None = None,
-        dtype: torch.dtype | None = None,
-    ):
-        super().__init__()
-
-        self.embed_dim = embed_dim
-        self.num_heads = num_heads
-        self.head_dim = embed_dim // num_heads
-        self.dropout = dropout
-
-        if embed_dim % num_heads != 0:
-            raise ValueError(
-                f"embed_dim ({embed_dim}) must be divisible by num_heads ({num_heads})"
-            )
-
-        # Fused QKV projection for better memory efficiency
-        self.qkv_proj = nn.Linear(embed_dim, 3 * embed_dim, bias=bias, device=device, dtype=dtype)
-
-        # Output projection
-        self.out_proj = nn.Linear(embed_dim, embed_dim, bias=bias, device=device, dtype=dtype)
-
-        # Dropout layer
-        if dropout > 0.0:
-            self.dropout_layer = nn.Dropout(dropout)
-        else:
-            self.dropout_layer = None
-
-        # Initialize weights
-        self._reset_parameters()
-
-    def _reset_parameters(self):
-        """Initialize projection weights"""
-        nn.init.xavier_uniform_(self.qkv_proj.weight)
-        if self.qkv_proj.bias is not None:
-            nn.init.constant_(self.qkv_proj.bias, 0.0)
-
-        nn.init.xavier_uniform_(self.out_proj.weight)
-        if self.out_proj.bias is not None:
-            nn.init.constant_(self.out_proj.bias, 0.0)
-
-    def forward(self, x: Tensor) -> tuple[Tensor, Tensor, Tensor]:
-        """
-        Forward pass with fused QKV projection.
-
-        Args:
-            x: Input tensor [batch, seq_len, embed_dim]
-
-        Returns:
-            q, k, v: Query, key, value tensors [batch, seq_len, num_heads, head_dim]
-        """
-        batch, seq_len, embed_dim = x.shape
-
-        # Fused QKV projection
-        qkv = self.qkv_proj(x)  # [batch, seq_len, 3 * embed_dim]
-
-        # Split and reshape for multi-head attention
-        qkv = qkv.view(batch, seq_len, 3, self.num_heads, self.head_dim)
-        q, k, v = qkv.unbind(dim=2)  # Each: [batch, seq_len, num_heads, head_dim]
-
-        return q, k, v
-
-    def project_output(self, attention_output: Tensor) -> Tensor:
-        """
-        Project attention output back to embedding dimension.
-
-        Args:
-            attention_output: [batch, seq_len, num_heads, head_dim]
-
-        Returns:
-            output: [batch, seq_len, embed_dim]
-        """
-        batch, seq_len, num_heads, head_dim = attention_output.shape
-
-        # Reshape to [batch, seq_len, embed_dim]
-        attention_output = attention_output.view(batch, seq_len, self.embed_dim)
-
-        # Apply output projection
-        output = self.out_proj(attention_output)
-
-        # Apply dropout if configured
-        if self.dropout_layer is not None:
-            output = self.dropout_layer(output)
-
-        return output
-
-
 class BlockSparseRingMultiheadDilatedAttention(nn.Module):
     """
-    Block-Sparse Ring Multihead Dilated Attention module.
+    Memory-efficient multihead block-sparse ring dilated attention.
 
-    A production-ready, drop-in replacement for nn.MultiheadAttention that provides
-    dramatic performance improvements through block-sparse patterns and Ring Attention.
-
-    Features:
-    - Compatible with nn.MultiheadAttention interface
-    - 10-100x speedup over standard attention
-    - 90-99% memory reduction
-    - Linear scaling to unlimited sequence lengths
-    - Advanced sparse pattern optimization
-    - Hardware-specific optimizations
-    - Comprehensive monitoring and debugging
+    Drop-in replacement for nn.MultiheadAttention with block-sparse optimization.
     """
 
     def __init__(
         self,
         embed_dim: int,
         num_heads: int,
-        segment_lengths: Sequence[int] = (2048, 4096, 8192),
-        dilation_rates: Sequence[int] = (1, 2, 4),
-        sparse_config: SparsePatternConfig | None = None,
+        segment_lengths: list[int],
+        dilation_rates: list[int],
         dropout: float = 0.0,
         bias: bool = True,
+        sparse_config: SparsePatternConfig | None = None,
         batch_first: bool = True,
-        use_adaptive_sparsity: bool = False,
-        ring_size: int = 1,
-        device: torch.device | None = None,
-        dtype: torch.dtype | None = None,
         **kwargs,
     ):
         """
-        Initialize Block-Sparse Ring Multihead Dilated Attention.
+        Initialize multihead block-sparse attention.
 
         Args:
-            embed_dim: Total dimension of the model
-            num_heads: Number of parallel attention heads
-            segment_lengths: Sequence of segment lengths for dilated attention
-            dilation_rates: Corresponding dilation rates
-            sparse_config: Configuration for sparsity patterns
+            embed_dim: Total embedding dimension
+            num_heads: Number of attention heads
+            segment_lengths: List of segment lengths for dilated attention
+            dilation_rates: List of dilation rates
             dropout: Dropout probability
-            bias: Whether to use bias in linear layers
-            batch_first: If True, input shape is [batch, seq, feature]
-            use_adaptive_sparsity: Whether to use learned sparsity patterns
-            ring_size: Number of devices in ring for distributed attention
-            device: Device to place the module on
-            dtype: Data type for the module
-            **kwargs: Additional arguments
+            bias: Whether to use bias in projections
+            sparse_config: Sparse pattern configuration
+            batch_first: Whether batch dimension is first
+            **kwargs: Additional arguments for BlockSparseRingDilatedAttention
         """
         super().__init__()
 
         self.embed_dim = embed_dim
         self.num_heads = num_heads
         self.head_dim = embed_dim // num_heads
-        self.dropout = dropout
         self.batch_first = batch_first
-        self.ring_size = ring_size
+        self.dropout = dropout
 
-        if embed_dim % num_heads != 0:
+        if self.head_dim * num_heads != embed_dim:
             raise ValueError(
                 f"embed_dim ({embed_dim}) must be divisible by num_heads ({num_heads})"
             )
 
-        # Sparse configuration
-        self.sparse_config = sparse_config or SparsePatternConfig()
-        self.use_adaptive_sparsity = use_adaptive_sparsity
+        # Projections
+        self.q_proj = nn.Linear(embed_dim, embed_dim, bias=bias)
+        self.k_proj = nn.Linear(embed_dim, embed_dim, bias=bias)
+        self.v_proj = nn.Linear(embed_dim, embed_dim, bias=bias)
+        self.out_proj = nn.Linear(embed_dim, embed_dim, bias=bias)
 
-        # Fused QKV projection
-        self.qkv_projection = FusedQKVProjection(
-            embed_dim=embed_dim,
-            num_heads=num_heads,
-            bias=bias,
-            dropout=dropout,
-            device=device,
-            dtype=dtype,
-        )
+        # Dropout
+        self.dropout_layer = nn.Dropout(dropout) if dropout > 0.0 else nn.Identity()
 
-        # Core sparse ring attention
-        self.attention = BlockSparseRingDilatedAttention(
+        # Sparse attention
+        self.sparse_attention = BlockSparseRingDilatedAttention(
             segment_lengths=segment_lengths,
             dilation_rates=dilation_rates,
-            sparse_config=self.sparse_config,
-            use_adaptive_sparsity=use_adaptive_sparsity,
-            ring_size=ring_size,
-            device=device,
+            sparse_config=sparse_config,
             **kwargs,
         )
 
-        # Layer normalization for stability (optional)
-        self.use_layer_norm = kwargs.get("use_layer_norm", False)
-        if self.use_layer_norm:
-            self.layer_norm = nn.LayerNorm(embed_dim, device=device, dtype=dtype)
-
-        # Performance monitoring
-        self.performance_monitor = PerformanceMonitor()
-
-        # Thread-safe execution
-        self._forward_lock = threading.Lock()
-
-        # Enable optimizations in base class
-        if hasattr(self.attention, "enable_memory_pool"):
-            self.attention.enable_memory_pool = True
-        if hasattr(self.attention, "enable_packed_comm"):
-            self.attention.enable_packed_comm = True
-        if hasattr(self.attention, "enable_hardware_opt"):
-            self.attention.enable_hardware_opt = True
+        # For compatibility with some implementations
+        self._always_return_tuple = False
 
     def forward(
         self,
         query: Tensor,
-        key: Tensor | None = None,
-        value: Tensor | None = None,
+        key: Tensor,
+        value: Tensor,
         key_padding_mask: Tensor | None = None,
-        need_weights: bool = True,
+        need_weights: bool = False,
         attn_mask: Tensor | None = None,
-        average_attn_weights: bool = True,
         is_causal: bool = False,
-    ) -> tuple[Tensor, Tensor | None]:
+        average_attn_weights: bool = True,
+    ) -> tuple[Tensor, dict[str, Any] | None]:
         """
-        Forward pass compatible with nn.MultiheadAttention interface.
+        Forward pass.
 
         Args:
-            query: Query tensor [seq_len, batch, embed_dim] or [batch, seq_len, embed_dim]
-            key: Key tensor (if None, uses query)
-            value: Value tensor (if None, uses query)
-            key_padding_mask: Padding mask for keys
+            query: Query tensor [batch, seq_len, embed_dim] if batch_first
+            key: Key tensor [batch, seq_len, embed_dim] if batch_first
+            value: Value tensor [batch, seq_len, embed_dim] if batch_first
+            key_padding_mask: Mask for padded keys [batch, seq_len]
             need_weights: Whether to return attention weights
-            attn_mask: Attention mask
-            average_attn_weights: Whether to average attention weights over heads
+            attn_mask: Attention mask [seq_len, seq_len] or [batch, seq_len, seq_len]
             is_causal: Whether to apply causal masking
+            average_attn_weights: Ignored (for compatibility)
 
         Returns:
-            output: Attention output
-            attn_weights: Attention weights (if need_weights=True)
+            output: Attention output [batch, seq_len, embed_dim]
+            attn_weights: Sparse attention weights dict if need_weights, else None
         """
-        # Handle input format (seq_first vs batch_first)
+        _ = average_attn_weights  # Ignored for compatibility
+        # Handle batch_first
         if not self.batch_first:
-            query = query.transpose(0, 1)  # [batch, seq_len, embed_dim]
-            if key is not None:
-                key = key.transpose(0, 1)
-            if value is not None:
-                value = value.transpose(0, 1)
+            query = query.transpose(0, 1)
+            key = key.transpose(0, 1)
+            value = value.transpose(0, 1)
 
-        # Use query for key/value if not provided (self-attention)
-        if key is None:
-            key = query
-        if value is None:
-            value = query
+        batch_size, seq_len, _ = query.shape
 
-        batch, seq_len, embed_dim = query.shape
+        # Projections
+        q = self.q_proj(query)
+        k = self.k_proj(key)
+        v = self.v_proj(value)
 
-        # Performance monitoring
-        with self.performance_monitor.time_forward():
-            # QKV projection
-            q, k, v = self.qkv_projection(query)
+        # Reshape for attention: [batch, seq_len, num_heads, head_dim]
+        q = q.view(batch_size, seq_len, self.num_heads, self.head_dim)
+        k = k.view(batch_size, seq_len, self.num_heads, self.head_dim)
+        v = v.view(batch_size, seq_len, self.num_heads, self.head_dim)
 
-            # Handle different key/value inputs
-            if key is not query:
-                _, k, _ = self.qkv_projection(key)
-            if value is not query:
-                _, _, v = self.qkv_projection(value)
-
-            # Apply attention mask if provided
-            if attn_mask is not None or key_padding_mask is not None:
-                q, k, v = self._apply_masks(q, k, v, attn_mask, key_padding_mask)
-
-            # Core sparse attention computation
-            attention_output, attention_weights = self.attention(
-                q, k, v, is_causal=is_causal, return_attention_weights=need_weights
-            )
-
-            # Project output
-            output = self.qkv_projection.project_output(attention_output)
-
-            # Apply layer normalization if enabled
-            if self.use_layer_norm:
-                output = self.layer_norm(output)
-
-        # Handle output format
-        if not self.batch_first:
-            output = output.transpose(0, 1)  # [seq_len, batch, embed_dim]
-
-        # Process attention weights for return
-        if need_weights and attention_weights is not None:
-            if average_attn_weights:
-                attention_weights = attention_weights.mean(dim=1)  # Average over heads
-            if not self.batch_first:
-                attention_weights = attention_weights.transpose(1, 2)
-        else:
-            attention_weights = None
-
-        return output, attention_weights
-
-    def _apply_masks(
-        self,
-        q: Tensor,
-        k: Tensor,
-        v: Tensor,
-        attn_mask: Tensor | None = None,
-        key_padding_mask: Tensor | None = None,
-    ) -> tuple[Tensor, Tensor, Tensor]:
-        """Apply attention and padding masks"""
-        # For simplicity, this implementation focuses on the core sparse attention
-        # Full mask support would require additional complexity in the sparse attention kernel
-        if attn_mask is not None:
-            warnings.warn(
-                "Attention masks are not fully supported in sparse attention mode. "
-                "Consider using is_causal=True for causal masking."
-            )
-
+        # Apply padding mask if provided
         if key_padding_mask is not None:
+            # key_padding_mask: [batch, seq_len] where True means padded
+            # Expand to [batch, seq_len, num_heads, 1]
+            mask = key_padding_mask.unsqueeze(2).unsqueeze(3)
+            k = k.masked_fill(mask, 0.0)
+            v = v.masked_fill(mask, 0.0)
+
+        # Apply attention mask if provided
+        if attn_mask is not None:
+            # This is more complex for sparse attention
+            # For now, we'll log a warning
+            import warnings
+
             warnings.warn(
-                "Key padding masks are not fully supported in sparse attention mode. "
-                "Consider preprocessing to remove padding tokens."
+                "attn_mask is not fully supported in BlockSparseRingMultiheadDilatedAttention. "
+                "Use is_causal=True for causal masking instead.",
+                stacklevel=2,
             )
 
-        return q, k, v
-
-    def set_sparsity_ratio(self, sparsity_ratio: float):
-        """Dynamically adjust sparsity ratio"""
-        self.attention.set_sparsity_ratio(sparsity_ratio)
-
-    def enable_adaptive_sparsity(self, enable: bool = True):
-        """Enable or disable adaptive sparsity learning"""
-        self.attention.enable_adaptive_sparsity(enable)
-
-    def get_performance_stats(self) -> dict[str, Any]:
-        """Get comprehensive performance statistics"""
-        attention_stats = self.attention.get_performance_stats()
-        monitor_stats = self.performance_monitor.get_stats()
-
-        return {
-            "attention_stats": attention_stats,
-            "execution_stats": monitor_stats,
-            "memory_info": self.get_memory_info(),
-        }
-
-    def get_memory_info(self) -> dict[str, Any]:
-        """Get memory usage information"""
-        base_info = self.attention.get_memory_info()
-
-        # Add multihead-specific information
-        multihead_info = {
-            "embed_dim": self.embed_dim,
-            "num_heads": self.num_heads,
-            "head_dim": self.head_dim,
-            "parameters": sum(p.numel() for p in self.parameters()),
-            "fused_qkv_projection": True,
-            "layer_norm_enabled": self.use_layer_norm,
-        }
-
-        base_info.update(multihead_info)
-        return base_info
-
-    def extra_repr(self) -> str:
-        """String representation of the module"""
-        return (
-            f"embed_dim={self.embed_dim}, num_heads={self.num_heads}, "
-            f"dropout={self.dropout}, sparse_ratio={self.sparse_config.sparsity_ratio}, "
-            f"pattern={self.sparse_config.pattern_type}, adaptive={self.use_adaptive_sparsity}"
-        )
-
-
-class PerformanceMonitor:
-    """Performance monitoring utility for tracking execution metrics"""
-
-    def __init__(self):
-        self.forward_times: list[float] = []
-        self.memory_usage: list[int] = []
-        self.total_forwards = 0
-        self._lock = threading.Lock()
-
-    def time_forward(self):
-        """Context manager for timing forward passes"""
-        return ForwardTimer(self)
-
-    def record_forward_time(self, time_ms: float):
-        """Record forward pass time"""
-        with self._lock:
-            self.forward_times.append(time_ms)
-            self.total_forwards += 1
-
-            # Keep only recent history
-            if len(self.forward_times) > 100:
-                self.forward_times.pop(0)
-
-    def record_memory_usage(self, memory_bytes: int):
-        """Record memory usage"""
-        with self._lock:
-            self.memory_usage.append(memory_bytes)
-
-            # Keep only recent history
-            if len(self.memory_usage) > 100:
-                self.memory_usage.pop(0)
-
-    def get_stats(self) -> dict[str, Any]:
-        """Get performance statistics"""
-        with self._lock:
-            if not self.forward_times:
-                return {
-                    "total_forwards": self.total_forwards,
-                    "avg_forward_time_ms": 0.0,
-                    "min_forward_time_ms": 0.0,
-                    "max_forward_time_ms": 0.0,
-                    "avg_memory_mb": 0.0,
-                }
-
-            return {
-                "total_forwards": self.total_forwards,
-                "avg_forward_time_ms": sum(self.forward_times) / len(self.forward_times),
-                "min_forward_time_ms": min(self.forward_times),
-                "max_forward_time_ms": max(self.forward_times),
-                "avg_memory_mb": (
-                    sum(self.memory_usage) / len(self.memory_usage) / 1024**2
-                    if self.memory_usage
-                    else 0.0
-                ),
-            }
-
-
-class ForwardTimer:
-    """Context manager for timing forward passes"""
-
-    def __init__(self, monitor: PerformanceMonitor):
-        self.monitor = monitor
-        self.start_time = None
-
-    def __enter__(self):
-        if torch.cuda.is_available():
-            torch.cuda.synchronize()
-        self.start_time = (
-            torch.cuda.Event(enable_timing=True) if torch.cuda.is_available() else None
-        )
-        if self.start_time:
-            self.start_time.record()
+        # Compute sparse attention
+        if need_weights:
+            attn_output, attn_weights = self.sparse_attention(
+                q, k, v, is_causal=is_causal, return_attention_weights=True
+            )
         else:
-            import time
+            attn_output = self.sparse_attention(
+                q, k, v, is_causal=is_causal, return_attention_weights=False
+            )
+            attn_weights = None
 
-            self.start_time = time.time()
-        return self
+        # Reshape output: [batch, seq_len, embed_dim]
+        attn_output = attn_output.view(batch_size, seq_len, self.embed_dim)
 
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        if torch.cuda.is_available():
-            end_time = torch.cuda.Event(enable_timing=True)
-            end_time.record()
-            torch.cuda.synchronize()
+        # Output projection
+        output = self.out_proj(attn_output)
+        output = self.dropout_layer(output)
 
-            if hasattr(self.start_time, "elapsed_time"):
-                elapsed_ms = self.start_time.elapsed_time(end_time)
-            else:
-                elapsed_ms = 0.0
-        else:
-            import time
+        # Handle batch_first for output
+        if not self.batch_first:
+            output = output.transpose(0, 1)
 
-            elapsed_ms = (time.time() - self.start_time) * 1000
-
-        self.monitor.record_forward_time(elapsed_ms)
-
-        # Record memory usage if CUDA available
-        if torch.cuda.is_available():
-            memory_bytes = torch.cuda.memory_allocated()
-            self.monitor.record_memory_usage(memory_bytes)
-
-
-# Convenience functions for easy creation
-def create_block_sparse_multihead_attention(
-    embed_dim: int,
-    num_heads: int,
-    sparsity_ratio: float = 0.25,
-    pattern_type: str = "dilated_sparse",
-    **kwargs,
-) -> BlockSparseRingMultiheadDilatedAttention:
-    """
-    Create a block-sparse multihead attention module with sensible defaults.
-
-    Args:
-        embed_dim: Model dimension
-        num_heads: Number of attention heads
-        sparsity_ratio: Fraction of blocks to compute (0.25 = 75% sparse)
-        pattern_type: Type of sparsity pattern ('local_window', 'dilated_sparse', 'global_local')
-        **kwargs: Additional arguments
-
-    Returns:
-        Configured block-sparse attention module
-    """
-    sparse_config = SparsePatternConfig(
-        pattern_type=pattern_type,
-        sparsity_ratio=sparsity_ratio,
-        block_size=kwargs.pop("block_size", 128),
-    )
-
-    return BlockSparseRingMultiheadDilatedAttention(
-        embed_dim=embed_dim, num_heads=num_heads, sparse_config=sparse_config, **kwargs
-    )
-
-
-def create_adaptive_sparse_multihead_attention(
-    embed_dim: int, num_heads: int, **kwargs
-) -> BlockSparseRingMultiheadDilatedAttention:
-    """
-    Create an adaptive sparse multihead attention module that learns optimal patterns.
-
-    Args:
-        embed_dim: Model dimension
-        num_heads: Number of attention heads
-        **kwargs: Additional arguments
-
-    Returns:
-        Configured adaptive sparse attention module
-    """
-    sparse_config = SparsePatternConfig(
-        pattern_type="adaptive",
-        sparsity_ratio=0.25,  # Starting point for adaptation
-        block_size=kwargs.pop("block_size", 128),
-    )
-
-    return BlockSparseRingMultiheadDilatedAttention(
-        embed_dim=embed_dim,
-        num_heads=num_heads,
-        sparse_config=sparse_config,
-        use_adaptive_sparsity=True,
-        **kwargs,
-    )
-
-
-# Export main classes and functions
-__all__ = [
-    "BlockSparseRingMultiheadDilatedAttention",
-    "FusedQKVProjection",
-    "PerformanceMonitor",
-    "create_adaptive_sparse_multihead_attention",
-    "create_block_sparse_multihead_attention",
-]
+        # Return format
+        if need_weights or self._always_return_tuple:
+            return output, attn_weights
+        return output
