--- docs/guides/block-sparse-v2-improvements.md
@@ -1,191 +1,0 @@
-# Block Sparse Ring Dilated Attention V2 - Memory Efficient Implementation
-
-## Overview
-
-The original BlockSparseRingDilatedAttention implementation had a critical flaw: despite being designed for sparsity, it actually used **more** memory than dense attention implementations. The V2 implementation fixes these issues and delivers true memory efficiency.
-
-## Key Problems with Original Implementation
-
-1. **Dense Attention Matrix Storage**: When `return_attention_weights=True`, allocated full `[batch, num_heads, seq_len, seq_len]` tensors
-2. **Memory Explosion**: For 64K sequences, this meant 64K × 64K = 4 billion elements!
-3. **Defeated Purpose**: The "sparse" implementation used more memory than dense versions
-
-## V2 Implementation Improvements
-
-### 1. True Sparse Storage
-
-Instead of materializing full attention matrices, V2 stores only active blocks:
-
-```python
-# Original: Dense matrix
-attention_weights = torch.zeros(batch, num_heads, seq_len, seq_len)  # Huge!
-
-# V2: Sparse format
-attention_weights = {
-    'block_indices': [(q_idx, k_idx), ...],  # Only active blocks
-    'block_values': [block_tensor, ...],     # Small block tensors
-    'shape': (batch, num_heads, seq_len, seq_len),
-    'block_size': 128
-}
-```
-
-### 2. Memory Efficiency Results
-
-From our tests with 8K sequences:
-
-| Implementation | Forward (no weights) | Forward (with weights) | Attention Storage |
-|----------------|---------------------|------------------------|-------------------|
-| Original       | 40.1 MB            | **OUT OF MEMORY**      | 2+ GB (attempted) |
-| V2             | 16.0 MB            | 16.0 MB                | 9.8 MB           |
-
-For larger sequences (4K test):
-- **Dense attention would use**: 256.0 MB
-- **V2 sparse attention uses**: 5.1 MB
-- **Memory savings**: 98.0%
-
-### 3. Block-Wise Processing
-
-V2 processes only active blocks, never materializing full matrices:
-
-```python
-for idx in range(num_active_blocks):
-    q_block = q_blocks[:, q_idx]
-    k_block = k_blocks[:, k_idx]
-    v_block = v_blocks[:, k_idx]
-    
-    # Process only this block
-    scores = torch.matmul(q_block, k_block.transpose(-2, -1))
-    # ... attention computation ...
-```
-
-### 4. Efficient Pattern Storage
-
-Patterns are stored as index lists, not dense matrices:
-
-```python
-# Instead of dense boolean matrix [num_blocks, num_blocks]
-# Store only active indices
-row_indices = [0, 0, 1, 1, 2, ...]
-col_indices = [0, 1, 0, 2, 1, ...]
-```
-
-## Usage
-
-### Basic Usage
-
-```python
-from dilated_attention_pytorch.block_sparse_ring_dilated_attention_v2 import (
-    BlockSparseRingDilatedAttentionV2,
-    SparsePatternConfig
-)
-
-# Configure sparsity
-sparse_config = SparsePatternConfig(
-    pattern_type='dilated_sparse',  # or 'local_window', 'global_local'
-    sparsity_ratio=0.1,             # Keep only 10% of blocks
-    block_size=128,                 # Size of each block
-)
-
-# Create module
-attention = BlockSparseRingDilatedAttentionV2(
-    segment_lengths=[2048, 4096, 8192],
-    dilation_rates=[1, 2, 4],
-    sparse_config=sparse_config,
-)
-
-# Forward pass
-output = attention(q, k, v)
-
-# Get sparse attention weights
-output, weights = attention(q, k, v, return_attention_weights=True)
-# weights is a dict with block indices and values
-```
-
-### Multihead Usage
-
-```python
-from dilated_attention_pytorch.block_sparse_ring_multihead_dilated_attention_v2 import (
-    BlockSparseRingMultiheadDilatedAttentionV2
-)
-
-multihead_attention = BlockSparseRingMultiheadDilatedAttentionV2(
-    embed_dim=768,
-    num_heads=12,
-    segment_lengths=[2048, 4096],
-    dilation_rates=[1, 2],
-    sparse_config=sparse_config,
-)
-
-output = multihead_attention(query, key, value)
-```
-
-## Sparsity Patterns
-
-### 1. Local Window
-Attends to nearby blocks within a window:
-```python
-sparse_config = SparsePatternConfig(
-    pattern_type='local_window',
-    local_window_size=512,  # Attend within 512 tokens
-    block_size=128,
-)
-```
-
-### 2. Dilated Sparse
-Hierarchical attention with multiple dilation rates:
-```python
-sparse_config = SparsePatternConfig(
-    pattern_type='dilated_sparse',
-    sparsity_ratio=0.1,  # Very sparse
-    block_size=128,
-)
-```
-
-### 3. Global + Local
-Combines global attention to first tokens with local windows:
-```python
-sparse_config = SparsePatternConfig(
-    pattern_type='global_local',
-    global_tokens=128,      # Attend to first 128 tokens globally
-    local_window_size=512,  # Plus local window
-    block_size=128,
-)
-```
-
-## Performance Guidelines
-
-1. **Block Size**: Larger blocks (128-256) are more efficient but less flexible
-2. **Sparsity Ratio**: Lower ratios (0.05-0.1) save more memory but may impact quality
-3. **Pattern Type**: 
-   - Use `local_window` for tasks with strong locality
-   - Use `dilated_sparse` for hierarchical/long-range dependencies
-   - Use `global_local` when some tokens need global context
-
-## Migration from Original
-
-The V2 implementation is a drop-in replacement with the same API, but returns attention weights in a different format:
-
-```python
-# Original returns dense tensor if requested
-output, weights = attention(q, k, v, return_attention_weights=True)
-# weights.shape = [batch, num_heads, seq_len, seq_len]  # HUGE!
-
-# V2 returns sparse dict
-output, weights = attention_v2(q, k, v, return_attention_weights=True)
-# weights = {
-#     'block_indices': list of (q_idx, k_idx) tuples,
-#     'block_values': list of attention blocks,
-#     'shape': original attention shape,
-#     'block_size': block size used
-# }
-```
-
-## Conclusion
-
-The V2 implementation delivers on the promise of sparse attention:
-- **98%+ memory savings** for attention weights with 95% sparsity
-- **No OOM errors** on long sequences
-- **True O(n)** memory complexity with sparsity
-- **Efficient block-wise computation**
-
-This makes it practical to process much longer sequences that would be impossible with the original implementation.
